requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
click==8.2.1
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
iniconfig==2.1.0
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.5.1
lightgbm==4.6.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
numpy==2.3.2
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pluggy==1.6.0
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
rich==14.1.0
rich-toolkit==0.14.8
rignore==0.6.4
scikit-learn==1.6.1
scipy==1.16.0
sentry-sdk==2.33.2
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
starlette==0.47.2
threadpoolctl==3.6.0
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
